Update _main.gsc
In this modification, i (the index of the player in the level.players array) is passed along with the “options|main” string. You’ll then need to modify the functions that use level.input["players"] to parse and use this index to identify the correct player.

Remember, modifications like this will require corresponding changes in other parts of the script where these menu actions are processed. Testing and debugging will be essential to ensure everything works as intended.